+++ osquery/tables/specs/utility/yara.table
@@ -1,0 +1,10 @@
+table_name("yara")
+description("Track YARA matches for files or PIDs.")
+schema([
+    Column("path", TEXT, "The path scanned"),
+    Column("matches", TEXT, "List of YARA matches"),
+    Column("count", INTEGER, "Number of YARA matches"),
+    Column("sig_group", TEXT, "Signature group used"),
+    Column("sigfile", TEXT, "Signature file used"),
+])
+implementation("yara@genYara")
